Added --arch to specify the architecture(s) to export
For example, this command will select two binaries for each component,
one that will run on an i586 machine and one that will run on an x86_64
machine:

  0export --arch Linux-i586 --arch Linux-x86_64 setup.sh http://...

If a single binary is the best option for both then it will
only be included in the bundle once. When the bundle is installed, only
the required versions will be copied into the machine's cache.