Install the distribution's zeroinstall-injector package if required
If the user asks to add the program to their Applications menu then
we need to make sure 0launch is in $PATH.

So, if 0launch isn't in $PATH then install the distribution package if
we have gnome-terminal, sudo and apt-get. This works on Ubuntu, but will
need modifications for other systems.

If that fails, tell the user to install it manually.